Refactor tests structure, include tests in dist tarball. Fixes #212
