+++ MANIFEST.in
@@ -2,3 +2,7 @@ include LICENSE
 include README.md
 include requirements.txt
 recursive-include inputs *
+recursive-include tests *.py
+include tests/requirements.txt
+include tests/mock_jsons.sh
+recursive-include tests/mock_jsons/ *
